requirements.txt
pandas
python-telegram-bot
mysql-connector-python